server/old/requirements.txt
fastapi[all]
requests
pillow